+++ changelog.md
@@ -1,4 +1,15 @@
 
+## 0.0.4 - 2015.08.21
+
+- added access tokens
+- added .env conf
+
+
+## 0.0.3 - 2015.08.21
+
+- switched to loose schemas
+
+
 ## 0.0.2 - 2015.08.19
 
 - added mongo-express
@@ -8,4 +19,4 @@
 
 ## 0.0.1 - 2015.08.09
 
-init
+- initial version